fixNotif
