+++ mobile_app/lib/main.dart
@@ -21,15 +21,6 @@ Future main() async {
   OneSignal.Debug.setLogLevel(OSLogLevel.verbose);
   OneSignal.initialize("5126ff04-dfdf-4525-95fd-c6cf3c630acd");
   OneSignal.Notifications.requestPermission(true);
-  final SharedPreferences sp = await SharedPreferences.getInstance();
-  var apiUrl = dotenv.env['API_HTTPS_URL'] ?? "";
-  var jwt = sp.getString("jwt") ?? "";
-  OneSignal.login(jwt);
-  var userId = await OneSignal.User.getExternalId();
-  print("EXTERNAL ID : ${userId}");
-  var route = Uri.https(apiUrl, "/api/setOneSignalId");
-  var apiRes = await http.post(route,
-      body: {'onesignal_id': "userId"}, headers: {'Authorization': jwt});
   runApp(const MyApp());
 }
